app/requirements.txt
google-cloud-pubsub
confluent-kafka
numpy